@@ -1,7 +1,7 @@
 {
   "name": "angular-electron",
-  "version": "12.0.1",
-  "description": "Angular 15 with Electron 24 (Typescript + SASS + Hot Reload)",
+  "version": "13.0.0",
+  "description": "Angular 16 with Electron 25 (Typescript + SASS + Hot Reload)",
   "homepage": "https://github.com/maximegris/angular-electron",
   "author": {
     "name": "Maxime GRIS",
@@ -9,9 +9,9 @@
   },
   "keywords": [
     "angular",
-    "angular 15",
+    "angular 16",
     "electron",
-    "electron 24",
+    "electron 25",
     "nodejs",
     "typescript",
     "playwright",
@@ -45,44 +45,44 @@
     "lint": "ng lint"
   },
   "dependencies": {
-    "@angular/common": "15.2.9",
-    "@angular/compiler": "15.2.9",
-    "@angular/core": "15.2.9",
-    "@angular/forms": "15.2.9",
-    "@angular/language-service": "15.2.9",
-    "@angular/platform-browser": "15.2.9",
-    "@angular/platform-browser-dynamic": "15.2.9",
-    "@angular/router": "15.2.9",
+    "@angular/common": "16.1.3",
+    "@angular/compiler": "16.1.3",
+    "@angular/core": "16.1.3",
+    "@angular/forms": "16.1.3",
+    "@angular/language-service": "16.1.3",
+    "@angular/platform-browser": "16.1.3",
+    "@angular/platform-browser-dynamic": "16.1.3",
+    "@angular/router": "16.1.3",
     "rxjs": "7.8.1",
     "tslib": "^2.5.0",
-    "zone.js": "~0.12.0"
+    "zone.js": "~0.13.1"
   },
   "devDependencies": {
-    "@angular-builders/custom-webpack": "15.0.0",
-    "@angular-devkit/build-angular": "15.2.8",
-    "@angular-eslint/builder": "15.2.1",
-    "@angular-eslint/eslint-plugin": "15.2.1",
-    "@angular-eslint/eslint-plugin-template": "15.2.1",
-    "@angular-eslint/schematics": "15.2.1",
-    "@angular-eslint/template-parser": "15.2.1",
-    "@angular/cli": "15.2.8",
-    "@angular/compiler-cli": "15.2.9",
-    "@ngx-translate/core": "14.0.0",
-    "@ngx-translate/http-loader": "7.0.0",
-    "@playwright/test": "1.33.0",
+    "@angular-builders/custom-webpack": "16.0.0",
+    "@angular-devkit/build-angular": "16.1.3",
+    "@angular-eslint/builder": "16.0.3",
+    "@angular-eslint/eslint-plugin": "16.0.3",
+    "@angular-eslint/eslint-plugin-template": "16.0.3",
+    "@angular-eslint/schematics": "16.0.3",
+    "@angular-eslint/template-parser": "16.0.3",
+    "@angular/cli": "16.1.3",
+    "@angular/compiler-cli": "16.1.3",
+    "@ngx-translate/core": "15.0.0",
+    "@ngx-translate/http-loader": "8.0.0",
+    "@playwright/test": "1.35.1",
     "@types/jasmine": "4.3.1",
     "@types/jasminewd2": "2.0.10",
     "@types/node": "18.16.9",
-    "@typescript-eslint/eslint-plugin": "5.59.5",
-    "@typescript-eslint/parser": "5.59.5",
-    "conventional-changelog-cli": "2.2.2",
-    "electron": "24.3.0",
+    "@typescript-eslint/eslint-plugin": "5.60.1",
+    "@typescript-eslint/parser": "5.60.1",
+    "conventional-changelog-cli": "3.0.0",
+    "electron": "25.2.0",
     "electron-builder": "24.4.0",
     "electron-debug": "3.2.0",
     "electron-reloader": "1.2.3",
-    "eslint": "8.40.0",
+    "eslint": "8.43.0",
     "eslint-plugin-import": "2.27.5",
-    "eslint-plugin-jsdoc": "44.2.3",
+    "eslint-plugin-jsdoc": "46.4.2",
     "eslint-plugin-prefer-arrow": "1.2.3",
     "jasmine-core": "4.6.0",
     "jasmine-spec-reporter": "7.0.0",
@@ -93,16 +93,16 @@
     "karma-jasmine-html-reporter": "2.0.0",
     "node-polyfill-webpack-plugin": "2.0.1",
     "npm-run-all": "4.1.5",
-    "playwright": "1.33.0",
+    "playwright": "1.35.1",
     "ts-node": "10.9.1",
     "typescript": "~4.9.5",
     "wait-on": "7.0.1",
     "webdriver-manager": "12.1.9"
   },
   "engines": {
-    "node": ">= 14.20.0 || >= 16.13.0 || >= 18.10.0"
+    "node": ">= 16.14.0 || >= 18.10.0"
   },
   "browserslist": [
-    "chrome 112"
+    "chrome 114"
   ]
 }
