@@ -1,8 +1,8 @@
 import { BrowserContext, ElectronApplication, Page, _electron as electron } from 'playwright';
 import { test, expect } from '@playwright/test';
-const PATH = require('path');
+import * as PATH from 'path';
 
-test.describe('Check Home Page', async () => {
+test.describe('Check Home Page', () => {
   let app: ElectronApplication;
   let firstWindow: Page;
   let context: BrowserContext;
@@ -48,7 +48,7 @@ test.describe('Check Home Page', async () => {
 
   test('Check title', async () => {
     const elem = await firstWindow.$('app-home h1');
-    const text = await elem.innerText();
+    const text = elem ? await elem.innerText() : null;
     expect(text).toBe('App works !');
   });
 
