@@ -18,12 +18,12 @@ export class ElectronService {
   constructor() {
     // Conditional imports
     if (this.isElectron) {
-      this.ipcRenderer = window.require('electron').ipcRenderer;
-      this.webFrame = window.require('electron').webFrame;
+      this.ipcRenderer = (window as any).require('electron').ipcRenderer;
+      this.webFrame = (window as any).require('electron').webFrame;
 
-      this.fs = window.require('fs');
+      this.fs = (window as any).require('fs');
 
-      this.childProcess = window.require('child_process');
+      this.childProcess = (window as any).require('child_process');
       this.childProcess.exec('node -v', (error, stdout, stderr) => {
         if (error) {
           console.error(`error: ${error.message}`);
